requirements.txt
pytesseract
opencv-python
numpy
scikit-image
deskew
pdf2image
python-docx
PyYAML
yamlloader
tqdm
Pillow
beautifulsoup4
layoutparser